backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
aiomysql==0.2.0
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1
